Make layerNew a mode in layer
